add: meta save data

+++ Assets/SoaringHorse/Scripts/Core/Services/ProgressSyncService.cs
@@ -10,7 +10,7 @@ public class ProgressSyncService
     public float ReadScore() =>
         _progress.Score;
 
-    public void SyncScore(float value) =>
+    public void SyncScore(int value) =>
         _progress.AddScore(value);
 
     public int ReadLifes() =>
