+++ Assets/SoaringHorse/Scripts/Data/Save/RunSaveData.cs
@@ -1,14 +1,14 @@
 using System;
-using UnityEngine;
+using UnityEngine.SocialPlatforms.Impl;
 
 [Serializable]
-public class RunSaveData 
+public class RunSaveData
 {
     public const int CurrentVersion = 1;
 
     public int version = CurrentVersion;
 
-    public float score = 0;
+    public int score = 0;
     public int lifes = 3;
     public float speed = 4;
     public float runTime = 0;
@@ -22,7 +22,7 @@ public class RunSaveData
         if (version <= 0) version = CurrentVersion;
         if (score < 0) score = 0;
         if (lifes < 0) lifes = 0;
-        if(speed < 4) speed = 4;
-        if(runTime < 0) runTime = 0;
+        if (speed < 4) speed = 4;
+        if (runTime < 0) runTime = 0;
     }
 }
